requirements.txt
pandas>=1.5.0
beautifulsoup4>=4.11.0
google-generativeai>=0.3.0
lxml>=4.9.0
requests>=2.28.0
html5lib>=1.1 